a2a_tutorial/beeai-a2a-server/requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.0
requests>=2.32.0
python-dotenv>=1.0.1
sse-starlette>=2.1.3
